Add tests and benchmark for HuffdicDecompressor, fix invalid shift
Extend test coverage to most parts of the HuffDic decompressor. This
requires a semantically correct dictionary.

Create a trivial dictionary for testing the HuffDic decompressor. The
dictionary entries are all for 8 bit symbols, and return an entry value
identical to the input symbol value. The data dictionary maps each entry
to a slice of lenght 1, containing a character value identical to the
offset.

A small dictionary (with only one dictionary containing all data
slices) also exposes a bug in the calculation of the dictionary number.
For 32 bit integers, valid shifts must be strictly less than 32, and
shifting by entry_bits == 32 is equivalent to `r >> (32 % 32)`.