Handle trailing data entries correctly
Text records may contain extra auxiliary data which should not be fed
to the decompressor.

The existence of such data is signalled by the `extraflags` header field,
and each set bit signals the corresponding extra data which will be
present in all text records.

The entries can be decoded (or removed) by reading the record from the
back. When an entry is present, its size will be at the very end of the
record, preceded by the actual data.

BUG: 475975
CCBUG: 482420
CCBUG: 489275