Reduce function call overhead, replace QByteArray buffer with std::vector
QByteArray::insert (as called by operator+=) is a function call, with
relevant overhead per call. std::vector::insert on the other hand is
completely inlined (as long as there is no need to grow the capacity).

For the typical compression ratio of 4 to 10 for Mobipocket XML content,
insert is called 400 to 1000 times per decompress() call.

The additional overhead of copying the data to the final QByteArray is
insignificant, and comparable to growing the QByteArray once. Growing
happened quite frequently, as the initial reservation is typically
not sufficient (compression ration is often much higher than 2). Instead
of estimating the compression ratio, use the input size of the compressor
(which equals the output size of the decompressor), which is 4096
(see maxRecordSize).

This reduces the instruction count from 22M to 16M for the dictionary
from BR 475975, with a similar reduction in runtime 	(1.77 vs 1.43
seconds).