Add missing bounds checks for getImage(), clean up, extend tests
The API allows to pass in negative indexes and indexes past the record
count. Also, the thumbnailIndex EXTH value can be out-of-bounds.

Verify the index value, and also check for overflow of image index +
firstImageRecord. An intermediate overflow can not be detected by
the PDB code so do it in the Document::image() code.

Remove some duplicated code.