Fix invalid QStringDecoder instantiation for cp1252
According to `QStringConverter::encodingForName(name)` documentation:

"If the name is not the name of a codec listed in the Encoding
enumeration, std::nullopt is returned. Such a name may, none the less,
be accepted by the QStringConverter constructor ...".

Notably, CP1252 is only accepted by the latter. Also change the
identifier to the standard one, "windows-1252".