Remove bogus fallback to random image for thumbnail
Images in Mobipocket files may appear in a random order not related to
the document order.

Until commit d1f0a3abc7ac ("Fix wrong calculation of thumbnail index.)
the thumbnail index was wrong in most cases - either default 0 without
an ThumbnailIndex EXTH field, or 12 (misinterpreting the TLV length
value). I.e. previously a fairly random image was returned, which
by chance sometimes matched the thumbnail or cover image.

After the fix a random image (index 0) was only returned when no thumbnail
was specified in the header.

Use the cover image as a fallback (which is typically a larger version
of the thumbnail), and remove the fallback to a random image.