Use QTextStream instead of qDebug()
If QDebug actually shows output depends on qtlogging.ini or enviroment
variables, and a tool named *dump not dumping the contents of a file is
at least surprising. Furthermore it makes the dump_meta and dump_text
tests fail if debug logging is filtered out.
The downside is that adding spaces and end-of-line characters doesn't
happen automatically.